types: make interfaces package strict-clean and remove it from the mypy ledger
Add a `from typing import Any` import to the BCI encoder, parameterise the EEG/
signal `np.ndarray` annotations with `np.ndarray[Any, Any]`, and type the
normalisation, legacy-scaling and firing-rate returns through annotated locals
(and the CCW frequency mapping through `float`). Drop the `interfaces.*` glob
from the mypy override ledger; the whole tree stays strict-clean.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/sc_neurocore/interfaces/ccw_bridge.py
@@ -116,7 +116,7 @@ class CCWBridge:
             Frequency in Hz mapped from bitstream probability
         """
         prob = np.mean(bitstream)
-        return freq_min + prob * (freq_max - freq_min)
+        return float(freq_min + prob * (freq_max - freq_min))
 
     def scpn_metrics_to_ccw(self, metrics: Dict[str, float]) -> Dict[str, float]:
         """
